+++ src/pygama/math/functions/__init__.py
