+++ src/pygama/math/functions/gauss_on_background.py
@@ -3,13 +3,14 @@ Gaussian distributions on linear and uniform backgrounds for pygama
 """
 import numpy as np
 
-from pygama.math._distributions.gauss import gauss_norm
+from pygama.math.functions.gauss import gauss_norm
 
 
 def gauss_uniform(x, n_sig, mu, sigma, n_bkg, components = False):
     """
     define a gaussian signal on a uniform background,
     args: n_sig mu, sigma for the signal and n_bkg for the background
+    TO DO: candidate for replacement by gauss_poly
     """
     if components==False:
         return 1/(np.nanmax(x)-np.nanmin(x)) * n_bkg + n_sig * gauss_norm(x,mu,sigma)
@@ -21,6 +22,7 @@ def gauss_linear(x, n_sig, mu, sigma, n_bkg, b, m, components=False):
     """
     gaussian signal + linear background function
     args: n_sig mu, sigma for the signal and n_bkg,b,m for the background
+    TO DO: candidate for replacement by gauss_poly
     """
     norm = (m/2 *np.nanmax(x)**2 + b*np.nanmax(x)) - (m/2 *np.nanmin(x)**2 + b*np.nanmin(x))
