+++ src/pygama/math/utils.py
@@ -3,50 +3,23 @@ pygama utility functions.
 """
 import sys
 
-import matplotlib.pyplot as plt
+
 import numpy as np
 import tqdm
-
-
-def get_dataset_from_cmdline(args, run_db, cal_db):
-    """
-    make it easier to call this from argparse:
-        arg("-ds", nargs='*', action="store", help="load runs for a DS")
-        arg("-r", "--run", nargs=1, help="load a single run")
-    """
-    from pygama import DataSet
-
-    if args["ds"]:
-        ds_lo = int(args["ds"][0])
-        try:
-            ds_hi = int(args["ds"][1])
-        except:
-            ds_hi = None
-        ds = DataSet(ds_lo, ds_hi, md=run_db, cal=cal_db, v=args["verbose"])
-
-    if args["run"]:
-        ds = DataSet(run=int(args["run"][0]), md=run_db, cal=cal_db,
-                     v=args["verbose"])
-    return ds
 
 
 def tqdm_range(start, stop, step=1, verbose=False, text=None, bar_length=20, unit=None):
     """
     Uses tqdm.trange which wraps around the python range and also has the option
     to display a progress
-
     For example:
     .. code-block :: python
-
         for start_row in range(0, tot_n_rows, buffer_len):
             ...
-
     Can be converted to the following
     .. code-block :: python
-
         for start_row in tqdm_range(0, tot_n_rows, buffer_len, verbose):
             ...
-
     Parameters
     ----------
     start : int
@@ -63,7 +36,7 @@ def tqdm_range(start, stop, step=1, verbose=False, text=None, bar_length=20, uni
         horizontal length of the bar in cursor spaces
     unit : str
         physical units to be displayed
-
+        
     Returns
     -------
     iterable : tqdm.trange
@@ -101,16 +74,6 @@ def sizeof_fmt(num, suffix='B'):
     return "{:.1f} {} {}".format(num, 'Y', suffix)
 
 
-def set_plot_style(style):
-    """
-    Choose a pygama plot style.
-    Current options: 'clint', 'root'
-    Or add your own [label].mpl file in the pygama directory!
-    """
-    path = __file__.rstrip('.utils.py')
-    plt.style.use(path+'/'+style+'.mpl')
-
-
 def get_par_names(func):
     """
     Return a list containing the names of the arguments of "func" other than the
@@ -119,19 +82,6 @@ def get_par_names(func):
     from scipy._lib._util import getargspec_no_self
     args, varargs, varkw, defaults = getargspec_no_self(func)
     return args[1:]
-
-
-def plot_func(func, pars, range=None, npx=None, **kwargs):
-    """
-    plot a function.  take care of the x-axis points automatically, or user can
-    specify via range and npx arguments.
-    """
-    if npx is None:
-        npx = 100
-    if range is None:
-        range = plt.xlim()
-    xvals = np.linspace(range[0], range[1], npx)
-    plt.plot(xvals, func(xvals, *pars), **kwargs)
 
 
 def get_formatted_stats(mean, sigma, ndigs=2):
@@ -165,28 +115,3 @@ def print_fit_results(pars, cov, func=None, title=None, pad=True):
         print(par_names[i], "=", mean, "+/-", sigma)
     if pad:
         print("")
-
-
-def tree_draw(tree, vars, tcut):
-    """
-    if you have to debase yourself and use ROOT, this is an easy
-    convenience function for quickly extracting data from TTrees.
-    TTree::Draw can only handle groups of 4 variables at a time,
-    but here we can put in as many as we want, and
-    return a list of numpy.ndarrays for each one
-    """
-    var_list = vars.split(":")
-    np_arrs = [[] for v in var_list]
-
-    for i in range(0, len(var_list), 4):
-
-        tmp_list = var_list[i:i + 4]
-        tmp_draw = ":".join(tmp_list)
-        n = tree.Draw(tmp_draw, tcut, "goff")
-
-        for j, var in enumerate(tmp_list):
-            # print(i, j, var, "getting V", j+1, "writing to np_arrs", i + j)
-            tmp = getattr(tree, f"GetV{j + 1}")()
-            np_arrs[i + j] = np.array([tmp[k] for k in range(n)])
-
-    return tuple(np_arrs)
