--- src/pygama/vis/mpl/root.mpl
@@ -1,35 +1,0 @@
-# https://matplotlib.org/users/customizing.html
-
-#figure.titlesize : small        # size of the figure title (Figure.suptitle())
-figure.dpi       : 120          # figure dots per inch
-
-lines.linewidth : 2
-lines.markersize : 10
-
-legend.fontsize : 14
-legend.handlelength : 1
-
-axes.titlesize : 17
-axes.labelsize : 17
-axes.grid : False   # display grid or not --> Doesn't work. Just type "g"
-
-xtick.direction : in
-ytick.direction : in
-xtick.labelsize : 14 # font size of tick labels
-ytick.labelsize : 14
-xtick.major.size     : 6      # major tick size in points
-xtick.minor.size     : 4      # minor tick size in points
-ytick.major.size     : 6
-ytick.minor.size     : 4
-xtick.minor.visible  : True   # turn minor ticks on by default
-ytick.minor.visible  : True
-
-# has to be a .ttf file in /Library/Fonts
-# font.family      : sans-serif
-# font.sans-serif : Helvetica Neue
-#font.sans-serif : 'Comic Sans'
-#font.style : italic
-# font.serif : Optima
-# font.family : Helvetica Neue Light
-# font.style : light
-#font.family : Optima
